+++ index.html
@@ -38,6 +38,6 @@
       <button id="hintBtn">Hint</button>
     </section>
   </main>
-  <script src="script.js"></script>
+  <script type="module" src="script.js"></script>
 </body>
 </html>
